Add `sumUpDownloadCounts` test
