+++ src/main/js/utils/sum-up-download-counts.js
@@ -1,0 +1,16 @@
+import $ from 'jquery';
+
+export function sumUpDownloadCounts(downloadData) {
+    var summedUpDownloads = {};
+    $.each(downloadData, function(packageName, packageDownloads) {
+        $.each(packageDownloads, function(date, packageDownloadsForDate) {
+            if (typeof summedUpDownloads[date] !== 'undefined') {
+                summedUpDownloads[date] = summedUpDownloads[date] + packageDownloadsForDate;
+            } else {
+                summedUpDownloads[date] = packageDownloadsForDate;
+            }
+        })
+    });
+
+    return {total: summedUpDownloads};
+}
